+++ app/src/main/java/com/jacpalberto/devcomms/events/EventsViewModel.kt
@@ -9,8 +9,10 @@ import com.jacpalberto.devcomms.data.DevCommsEvent
  * Created by Alberto Carrillo on 7/13/18.
  */
 class EventsViewModel : ViewModel() {
-    var fragmentTag: String = EventsByDateFragment.TAG
+    //TODO: inject model
     private var model = EventsModel()
+
+    var fragmentTag: String = EventsByDateFragment.TAG
     private var eventsLiveData: MutableLiveData<DataResponse<List<DevCommsEvent>>> = MutableLiveData()
     private var favoriteEventsLiveData: MutableLiveData<DataResponse<List<DevCommsEvent>>> = MutableLiveData()