+++ app/src/main/java/com/jacpalberto/devcomms/sponsors/SponsorsViewModel.kt
@@ -9,8 +9,10 @@ import com.jacpalberto.devcomms.data.Sponsor
  * Created by Alberto Carrillo on 8/13/18.
  */
 class SponsorsViewModel : ViewModel() {
-    private var sponsorsLiveData: MutableLiveData<DataResponse<List<Sponsor>>>? = null
+    //TODO: inject model
     private val model = SponsorsModel()
+
+    private var sponsorsLiveData: MutableLiveData<DataResponse<List<Sponsor>>>? = null
 
     fun fetchSponsors(): MutableLiveData<DataResponse<List<Sponsor>>>? {
         if (sponsorsLiveData == null) {